Refactored + packaged code
