+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.data.Storage;
+import duke.object.Task;
+import duke.object.TaskList;
+import duke.user.Ui;
+
 import java.io.IOException;
 
 public class AddEventCommand extends Command {
@@ -10,13 +17,13 @@ public class AddEventCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         ui.confirmEventAdded(tasks.addTask(new Task('E', desc, dateTime)), tasks);
         storage.saveTasks(tasks);
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
