+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,9 +1,17 @@
+package duke.command;
+
+import duke.data.Storage;
+import duke.exception.DukeException;
+import duke.object.Task;
+import duke.object.TaskList;
+import duke.user.Ui;
+
 import java.io.IOException;
 
 public class AddTodoCommand extends Command {
     String todo;
 
-    AddTodoCommand(String todo) {
+    public AddTodoCommand(String todo) {
         if (todo.equals("")) {
             throw new DukeException("    ☹ OOPS!!! The description of a todo cannot be empty.");
         }
@@ -11,13 +19,13 @@ public class AddTodoCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         ui.confirmTodoAdded(tasks.addTask(new Task('T', this.todo)), tasks);
         storage.saveTasks(tasks);
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
