+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.data.Storage;
+import duke.object.TaskList;
+import duke.user.Ui;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
@@ -8,13 +14,13 @@ public class DeleteCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         ui.confirmDelete(tasks.delete(idx));
         storage.saveTasks(tasks);
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
