+++ src/main/java/duke/command/ExitCommand.java
@@ -1,0 +1,18 @@
+package duke.command;
+
+import duke.data.Storage;
+import duke.object.TaskList;
+import duke.user.Ui;
+
+public class ExitCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.sayBye();
+    }
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
