+++ src/main/java/duke/command/FaultyCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.data.Storage;
+import duke.exception.DukeException;
+import duke.object.TaskList;
+import duke.user.Ui;
+
 import java.io.IOException;
 
 public class FaultyCommand extends Command {
@@ -8,12 +15,12 @@ public class FaultyCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
         ui.showError(e);
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 }
