+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 import java.util.InputMismatchException;
 
 public class DukeException extends InputMismatchException {
