+++ src/main/java/duke/object/Task.java
@@ -1,3 +1,5 @@
+package duke.object;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
