+++ src/main/java/duke/object/TaskList.java
@@ -1,3 +1,5 @@
+package duke.object;
+
 import java.util.ArrayList;
 
 public class TaskList {
